acviewertask: fix AcViewerHudDraw signature and drop the cast
Same arm64 callback ABI bug as MainTask::AepDrawCallback:
AcViewerHudDraw declared param 11 as int16_t rotation (2 bytes) instead
of int, so the stack-passed args shifted and corrupted context. Its p13
was already an int *; make rotation/blend/p14 match AepGroupDrawFn
exactly and register &AcViewerHudDraw without the reinterpret_cast. The
old comment claiming short was ABI-compatible was wrong: params 9+ are
on the stack, not in a register.

+++ Project/System/src/Task/AcViewerTask.h
@@ -68,10 +68,10 @@ void AcViewerHudDraw(int child,
                      int anchorY,
                      int color,
                      int alpha,
-                     int16_t rotation,
-                     int blend,
+                     int rotation,
+                     uint32_t blend,
                      int *p13,
-                     int p14,
+                     uint32_t p14,
                      void *context);
 
 class AcViewerTask : public C_TASK {
@@ -106,10 +106,10 @@ private:
                                 int anchorY,
                                 int color,
                                 int alpha,
-                                int16_t rotation,
-                                int blend,
+                                int rotation,
+                                uint32_t blend,
                                 int *p13,
-                                int p14,
+                                uint32_t p14,
                                 void *context);
 
     // The apply-settings bridge hook writes the option / seek fields below
